Created Function to Output Results
In this push I created a function that takes the output from the
Start-NcPerfPull and the environment information and places it into a
custom XML document that references a XSL for transformation with a
browser.

My thinking at this point is that I will be able to use these outputs to
be able to gather a trend so that people will be able to see how their
environment performance at different times throughout the day. A later
function can be built to compare these or to suck them all in and
reference them.