Fixed Error Display and protocol specification
In the input method there is a choice of what medium is having
performance degredation. This was binary anded against the csv file.
unfortunately it was not working due to the addition of the odd numbers.
Each number was rounded up to the next binary even number and the
results are accurate now.

The error display was looking at a count and not the standard deviation
between itterations. In Changing this I was able to find the difference
between the itterations in order to ellude to the change. Majority of
these values are not reset each itteration but are added. This provides
for more accurate results.